move config validation functions to root.go
